updating account

+++ micro/account/src/main/java/com/tagokoder/account/domain/port/in/GetAccountBalancesUseCase.java
@@ -1,10 +1,11 @@
 package com.tagokoder.account.domain.port.in;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface GetAccountBalancesUseCase {
 
-    record Result(UUID accountId, double ledger, double available, double hold) {}
+    record Result(UUID accountId, BigDecimal ledger, BigDecimal available, BigDecimal hold) {}
 
     Result get(UUID accountId);
 }
