+++ micro/account/src/main/java/com/tagokoder/account/domain/port/in/PatchAccountLimitsUseCase.java
@@ -1,11 +1,12 @@
 package com.tagokoder.account.domain.port.in;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PatchAccountLimitsUseCase {
 
-    record Command(UUID accountId, Double dailyOut, Double dailyIn) {}
-    record Result(UUID accountId, double dailyOut, double dailyIn) {}
+    record Command(UUID accountId, BigDecimal dailyOut, BigDecimal dailyIn) {}
+    record Result(UUID accountId, BigDecimal dailyOut, BigDecimal dailyIn) {}
 
     Result patch(Command command);
 }
