+++ micro/account/src/main/java/com/tagokoder/account/domain/port/in/ReleaseHoldUseCase.java
@@ -1,9 +1,10 @@
 package com.tagokoder.account.domain.port.in;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface ReleaseHoldUseCase {
-    record Command(UUID accountId, String currency, Double amount, String reasonOrNull) {}
-    record Result(boolean ok, Double newHold) {}
+    record Command(UUID accountId, String currency, BigDecimal amount, String reasonOrNull) {}
+    record Result(boolean ok, BigDecimal newHold) {}
     Result release(Command command);
 }
